+++ src/Ampersand/Prototype/GenFrontend.hs
@@ -1,5 +1,5 @@
 {-# LANGUAGE DeriveDataTypeable,OverloadedStrings #-}
-module Ampersand.Prototype.GenFrontend (doGenFrontend, clearTemplateDirs) where
+module Ampersand.Prototype.GenFrontend (doGenFrontend) where
 
 import           Ampersand.Basics
 import           Ampersand.Classes.Relational
@@ -9,14 +9,20 @@ import           Ampersand.FSpec.FSpec
 import           Ampersand.FSpec.ToFSpec.NormalForms
 import           Ampersand.Misc
 import           Ampersand.Prototype.ProtoUtil
+import           Codec.Archive.Zip
+import           Control.Exception
 import           Control.Monad
+import qualified Data.ByteString.Lazy  as BL
 import           Data.Data
 import           Data.List
 import           Data.Maybe
+import           Network.HTTP.Simple
 import           System.Directory
 import           System.FilePath
 import           Text.StringTemplate
 import           Text.StringTemplate.GenericStandard () -- only import instances
+
+
 
 {- TODO
 - Be more consistent with record selectors/pattern matching
@@ -51,18 +57,6 @@ This is considered editable iff the composition rel;relRef yields an editable re
 
 getTemplateDir :: FSpec -> String
 getTemplateDir fSpec = dirPrototype (getOpts fSpec) </> "templates"
-
--- Clear template dirs so the generator won't use lingering template files. 
--- (Needs to be called before statics are generated, otherwise the templates from statics/ZwolleFrontend/templates will get deleted)
--- TODO: refactor generate, so we can call generation of static files and generics.php from this module.
-clearTemplateDirs :: FSpec -> IO ()
-clearTemplateDirs fSpec = mapM_ emptyDir ["views", "controllers"]
-  where emptyDir path = 
-         do { let absPath = getTemplateDir fSpec </> path
-            ; dirExists <- doesDirectoryExist absPath
-            ; when dirExists $ -- dir may not exist if we haven't generated before
-               removeAllDirectoryFiles absPath
-            } -- Only remove files, withouth entering subdirectories, to prevent possible disasters with symbolic links.
         
 -- For useful info on the template language, see
 -- https://theantlrguy.atlassian.net/wiki/display/ST4/StringTemplate+cheat+sheet
@@ -72,22 +66,25 @@ clearTemplateDirs fSpec = mapM_ emptyDir ["views", "controllers"]
 
 doGenFrontend :: FSpec -> IO ()
 doGenFrontend fSpec =
- do { putStr "Generating frontend..\n" 
+ do { putStrLn "Generating frontend.."
+    ; downloadPrototypeFramework (getOpts fSpec)
     ; copyTemplates fSpec
     ; feInterfaces <- buildInterfaces fSpec
     ; genViewInterfaces fSpec feInterfaces
     ; genControllerInterfaces fSpec feInterfaces
     ; genRouteProvider fSpec feInterfaces
     ; copyCustomizations fSpec
-    ; deleteTemplateDir fSpec
-    ; putStrLn "Frontend generated.\n"
+    -- ; deleteTemplateDir fSpec -- don't delete template dir anymore, because it is required the next time the frontend is generated
+    ; putStrLn "Installing dependencies.."
+    ; installComposerLibs (getOpts fSpec)
+    ; putStrLn "Frontend generated."
     }
 
 copyTemplates :: FSpec -> IO ()
 copyTemplates fSpec =
  do { let adlSourceDir = takeDirectory $ fileName (getOpts fSpec)
           tempDir = adlSourceDir </> "templates"
-          toDir = (dirPrototype (getOpts fSpec)) </> "templates"
+          toDir = dirPrototype (getOpts fSpec) </> "templates"
     ; tempDirExists <- doesDirectoryExist tempDir
     ; if tempDirExists then
         do { verboseLn (getOpts fSpec) $ "Copying project specific templates from " ++ tempDir ++ " -> " ++ toDir
@@ -111,11 +108,10 @@ copyCustomizations fSpec =
         if sourceDirExists then
           do verboseLn opts $ "Copying customizations from " ++ sourceDir ++ " -> " ++ targetDir
              copyDirRecursively sourceDir targetDir opts -- recursively copy all customizations
-        else
-          do verboseLn opts $ "No customizations (there is no directory " ++ sourceDir ++ ")"
+        else verboseLn opts $ "No customizations (there is no directory " ++ sourceDir ++ ")"
 
-deleteTemplateDir :: FSpec -> IO ()
-deleteTemplateDir fSpec = removeDirectoryRecursive $ dirPrototype (getOpts fSpec) </> "templates"
+-- deleteTemplateDir :: FSpec -> IO ()
+-- deleteTemplateDir fSpec = removeDirectoryRecursive $ dirPrototype (getOpts fSpec) </> "templates"
 
 ------ Build intermediate data structure
 -- NOTE: _ disables 'not used' warning for fields
@@ -445,3 +441,69 @@ renderTemplate (Template template absPath) setAttrs =
              ([], attrs@(_:_), _)        -> templateError $ "The following attributes are expected by the template, but not supplied: " ++ show attrs
              ([], [], ts@(_:_)) -> templateError $ "Missing invoked templates: " ++ show ts -- should not happen as we don't invoke templates
   where templateError msg = error $ "\n\n*** TEMPLATE ERROR in:\n" ++ absPath ++ "\n\n" ++ msg
+
+
+
+
+downloadPrototypeFramework :: Options -> IO ()
+downloadPrototypeFramework opts = 
+  (do 
+    x <- allowExtraction
+    when x $ do 
+      when (forceReinstallFramework opts) destroyDestinationDir
+      verboseLn opts "Start downloading frontend framework."
+      response <- 
+        parseRequest ("https://github.com/AmpersandTarski/Prototype/archive/"++zwolleVersion opts++".zip") >>=
+        httpBS  
+      let archive = removeTopLevelFolder 
+                  . toArchive 
+                  . BL.fromStrict 
+                  . getResponseBody $ response
+      verboseLn opts "Start extraction of frontend framework."
+      let zipoptions = 
+              [OptVerbose | verboseP opts]
+            ++ [OptDestination destination]
+      extractFilesFromArchive zipoptions archive
+      writeFile (destination </> ".prototypeSHA")
+                (show . zComment $ archive)
+  ) `catch` \err ->  -- git failed to execute
+         exitWith . FailedToInstallPrototypeFramework $
+            [ "Error encountered during installation of prototype framework:"
+            , show (err :: SomeException)
+            ]
+            
+  where
+    destination = dirPrototype opts
+    destroyDestinationDir :: IO ()
+    destroyDestinationDir = removeDirectoryRecursive destination
+    removeTopLevelFolder :: Archive -> Archive
+    removeTopLevelFolder archive = 
+       archive{zEntries = mapMaybe removeTopLevelPath . zEntries $ archive}
+      where
+        removeTopLevelPath :: Entry -> Maybe Entry
+        removeTopLevelPath entry = 
+            case tail . splitPath . eRelativePath $ entry of
+              [] -> Nothing
+              xs -> Just entry{eRelativePath = joinPath xs}
+
+    allowExtraction :: IO Bool
+    allowExtraction = do
+      pathExist <- doesPathExist destination
+      destIsDirectory <- doesDirectoryExist destination 
+      if pathExist
+      then 
+          if destIsDirectory
+          then do 
+            dirContents <- listDirectory destination
+            unless (null dirContents)
+                   (verboseLn opts $
+                         "Didn't install prototype framework, because\n"
+                      ++ "  "++destination++" isn't empty.")
+            return (null dirContents)
+          else do 
+             verboseLn opts $
+                       "Didn't install prototype framework, because\n"
+                    ++ "  "++destination++" isn't a directory."
+             return False
+      else return True
+