+++ ReleaseNotes.md
@@ -9,6 +9,7 @@
 * Upgrade to pandoc vs 2.0
 * [Issue #482](https://github.com/AmpersandTarski/Ampersand/issues/482) Using checksum from generator to check if Ampersand model files are changed
 * [Issue #792](https://github.com/AmpersandTarski/Ampersand/issues/792) Add possibility to extend TXT in interfaces
+* [Issue #801](https://github.com/AmpersandTarski/Ampersand/issues/801) Fix a problem with empty values in .xlsx files during compilation of an .adl file
 
 ## v3.10.1 (8 june 2018)